WIP: fetch github repo from haskell
