+++ waspc/cli/src/Wasp/Cli/Command/CreateNewProject/StarterTemplates/Github.hs
@@ -1,10 +1,6 @@
 module Wasp.Cli.Command.CreateNewProject.StarterTemplates.Github where
 
-type GithubRepo = (GithubRepoOwner, GithubRepoName)
-
-type GithubRepoOwner = String
-
-type GithubRepoName = String
+import Wasp.Cli.GithubRepo (GithubRepo)
 
 starterTemplateGithubRepo :: GithubRepo
 starterTemplateGithubRepo = ("wasp-lang", "starters")
